requirements.txt
streamlit>=1.32.0
mysql-connector-python>=8.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
